src/requirements.txt
streamlit
chromadb
requests
sentence-transformers
pyyaml

